fix: Update to new generator api
