+++ packages/graphback-runtime-mongodb/src/index.ts
@@ -1,4 +1,4 @@
 
 export * from "./MongoDBDataProvider"
 export * from "./OffixMongoDBDataProvider"
-export * from "./runtimeContext"
+export * from "./createMongoServices"
