+++ packages/graphback-runtime/src/GraphbackModel.ts
@@ -1,0 +1,18 @@
+
+/**
+ * Interface represents graphback models that are generated by default resolver generator
+ */
+export interface GraphbackModel {
+    name: string,
+    pubSub: PubSubConfig
+}
+
+/**
+ * Configuration for the publish subscribe model
+ */
+export interface PubSubConfig {
+    publishCreate?: boolean;
+    publishUpdate?: boolean;
+    publishDelete?: boolean;
+}
+
