+++ packages/graphback-runtime/src/index.ts
@@ -2,6 +2,7 @@
 // DB
 export * from "./data/GraphbackDataProvider"
 export * from "./data/NoDataError"
+
 // Service
 export * from "./service/CRUDService"
 export * from "./service/GraphbackCRUDService"
@@ -10,3 +11,5 @@ export * from "./service/PubSubConfig"
 // Runtime
 export * from "./resolvers/LayeredRuntimeResolverCreator"
 export * from "./resolvers/RuntimeResolversDefinition"
+
+export * from "./GraphbackModel"
