+++ packages/graphback-runtime/src/service/CRUDService.ts
@@ -5,7 +5,7 @@ import { PubSubEngine } from 'graphql-subscriptions';
 import { GraphbackDataProvider } from "../data/GraphbackDataProvider";
 import { defaultLogger, GraphbackMessageLogger } from '../utils/Logger';
 import { GraphbackCRUDService } from "./GraphbackCRUDService";
-import { PubSubConfig } from "./PubSubConfig"
+import { GraphbackPubSub } from "./PubSubConfig"
 import { subscriptionTopicMapping } from './subscriptionTopicMapping';
 
 /**
@@ -19,10 +19,10 @@ export class CRUDService<T = any> implements GraphbackCRUDService<T>  {
     private db: GraphbackDataProvider;
     private logger: GraphbackMessageLogger;
     private pubSub: PubSubEngine;
-    private publishConfig: PubSubConfig;
+    private publishConfig: GraphbackPubSub;
     private modelName: string;
 
-    public constructor(modelType: GraphQLObjectType, db: GraphbackDataProvider, subscriptionConfig: PubSubConfig, logger?: GraphbackMessageLogger) {
+    public constructor(modelType: GraphQLObjectType, db: GraphbackDataProvider, subscriptionConfig: GraphbackPubSub, logger?: GraphbackMessageLogger) {
         this.db = db;
         this.pubSub = subscriptionConfig.pubSub;
         this.logger = logger || defaultLogger;
